Move def of Layout to its own submodules
